pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-pde-stl-strel"
version = "0.7.9"
description = "Reproducible STL/STREL monitoring and training-time regularization for neural PDE surrogates"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [
    {name = "Arthur Hu"},
]
keywords = [
    "physics-informed neural networks", "signal temporal logic",
    "STREL", "formal verification", "PDE", "PINN",
    "runtime monitoring", "spatio-temporal logic", "robustness",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]
stl = ["rtamt>=0.3; python_version < \"3.12\""]
strel = ["moonlight>=0.3"]
frameworks = ["torch>=2.0", "neuromancer", "torchphysics"]
plot = ["matplotlib>=3.7", "pandas>=2.0"]
dev = ["pytest>=7.0", "ruff>=0.4"]
all = [
    "torch>=2.0",
    "rtamt>=0.3; python_version < \"3.12\"",
    "moonlight>=0.3",
    "neuromancer",
    "torchphysics",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
neural-pde-stl-strel = "neural_pde_stl_strel.__main__:main"

[project.urls]
Repository = "https://github.com/arthurhu9413/neural-pde-stl-strel"

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
neural_pde_stl_strel = ["py.typed"]

[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "__pycache__/*"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
# Keep the default Ruff invocation aligned with the repository's CI/Makefile
# fatal-error gate. Broader style modernization is handled opportunistically.
select = ["E9", "F63", "F7", "F82"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_torch: marks tests that require PyTorch",
    "requires_rtamt: marks tests that require RTAMT",
    "requires_moonlight: marks tests that require MoonLight",
]
